pyproject.toml
[project]
name = "creatify-django-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
]
